Working example with jmfurlott's `react-styleguide`
This is not the perfect tool for practicing a Styleguide Driven Development, as
this tool does not generate a styleguide automatically.
Ideally, it would be some sort of documentation based generator, that works well
with the current application's building system.

Inspiration:

* https://github.com/pocotan001/react-styleguide-generator: this allows us to
document the component itself and the styleguide's component versions, with the
capabilities to exhibit different versions of the same component;

The reason I'm not using it is because it doesn't work with webpack, which is
  currently in our build system.
* https://github.com/sapegin/react-styleguidist: This one seems to have some
potentially nice features regarding component documentation and automatic
styleguide's component version generation. It doesn't allow to customize each
section's details, like section/subsection number, and/or order.
The problem seems to be being underdeveloped and a first attempt at using it
didn't work out well.